+++ 4.0/redmine/plugins/additionals/lib/additionals/wiki_macros/project_macro.rb
@@ -24,14 +24,14 @@ module Additionals
       macro :projects do |_obj, args|
         _args, options = extract_macro_options(args, :title, :with_create_issue)
         @projects = Additionals.load_projects
-        return '' if @projects.nil?
+        return if @projects.nil?
 
         @html_options = { class: 'external' }
         render partial: 'wiki/project_macros',
                formats: [:html],
-               locals:  { projects: @projects,
-                          list_title: options[:title],
-                          with_create_issue: options[:with_create_issue] }
+               locals: { projects: @projects,
+                         list_title: options[:title],
+                         with_create_issue: options[:with_create_issue] }
       end
     end
   end