+++ 4.0/redmine/plugins/redmine_lightbox2/assets/javascripts/lightbox.js
@@ -1,16 +1,26 @@
 $(document).ready(function() {
   // the file extension regex matching on supported image and pdf types
-  var extensionRegex = /\.(png|jpe?g|gif|pdf)$/i;
+  var extensionRegexImage = /\.(png|jpe?g|gif|bmp)$/i;
+  var extensionRegexAll = /\.(png|jpe?g|gif|bmp|pdf)$/i;
 
   // modify thumbnail links in wiki content -> add filename from ./img/@alt to url to support fancybox preview
   $("div.wiki a.thumbnail").attr('href', function(i, v){
-    return v.replace(/\/attachments\/(\d+)/g,'/attachments/download/$1') + '/' + $(this).children('img').attr('alt').replace(/(.*\.(png|jpe?g|gif))(\s\(.*\))?/gi,'$1');
+    return v.replace(/\/attachments\/(\d+)/g,'/attachments/download/$1') + '/' + $(this).children('img').attr('alt').replace(/(.*\.(png|jpe?g|gif|bmp))(\s\(.*\))?/gi,'$1');
   });
 
-  // modify thumbnails and magnifier links in journal details -> add filename to url to support fancybox preview
-  $("div.journal div.thumbnails a, div.journal ul.details li a:not([title])").attr('href', function(i, v){
-    if($(this).attr('href').match(extensionRegex)) {
-      return v.replace(/\/attachments\/(\d+)/g,'/attachments/download/$1');
+  // modify filename links in journal details -> add filename to url to support fancybox preview
+  $("div.journal ul.details li a:not([title])").attr('href', function(i, v){
+    if($(this).html().match(extensionRegexAll)) {
+      return v.replace(/\/attachments\/(\d+)/g,'/attachments/download/$1') + '/' + $(this).html();
+    } else {
+      return v;
+    }
+  });
+
+  // modify thumbnail links after journal details -> add filename to url to support fancybox preview
+  $("div.journal div.thumbnails a[title]").attr('href', function(i, v){
+    if($(this).attr('title').match(extensionRegexAll)) {
+      return v.replace(/\/attachments\/(\d+)/g,'/attachments/download/$1') + '/' + $(this).attr('title');
     } else {
       return v;
     }
@@ -18,7 +28,7 @@ $(document).ready(function() {
 
   // add a magnifier icon before download icon for images and pdf
   $("div.journal ul.details li a.icon-download").each(function(i, obj) {
-    if($(this).attr('href').match(extensionRegex)) {
+    if($(this).attr('href').match(extensionRegexAll)) {
       var icon = $(this).clone().attr('class', function(i, v){
         return v.replace(/-download/g,'-magnifier');
       });
@@ -32,12 +42,12 @@ $(document).ready(function() {
   });
 
   // #40 DMSF support: add class="thumbnail" to DMSF macro thumbnails
-  $("a[data-downloadurl][href^='/dmsf/files/'][href$='/view']").each(function(i, obj) {
+  $("div.wiki a[data-downloadurl][href^='/dmsf/files/'][href$='/view']").each(function(i, obj) {
     var filename = $(this).attr('data-downloadurl').split(':')[1];
     // Also support PDF preview in lightbox
     var isPdf = filename.match(/\.pdf$/i);
-    // Bugfix: only apply thumbnail class to image and pdf links
-    if(filename.match(extensionRegex)) {
+    // only apply thumbnail class to image and pdf links
+    if(filename.match(extensionRegexAll)) {
       $(this)
         .attr('class', 'thumbnail')
         .attr('data-fancybox-type', isPdf ? 'iframe' : 'image')
@@ -47,43 +57,62 @@ $(document).ready(function() {
     }
   });
 
-  // Add Fancybox to image links
-  $("div.attachments a.lightbox," +
-    "div.attachments a.lightbox_preview," +
-    "div.journal ul.details a[href$='.png']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.PNG']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.jpg']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.JPG']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.jpeg']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.JPEG']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.gif']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.GIF']:not(.icon-download)," +
-    "div.journal div.thumbnails a," +
-    "div.wiki a.thumbnail," +
-    "div.attachments a.swf," +
-    ".avatar a").fancybox({
-      prevEffect    : 'none',
-      nextEffect    : 'none',
-      openSpeed     : 300,
-      closeSpeed    : 150
+  // #53 DMSF support in issues: add class="lightbox" to DMSF thumbnails and preview links
+  $("div.attachments.dmsf_parent_container a[href^='/dmsf/files/'][href$='/view']").each(function(i, obj) {
+    // extract filename from attribute 'data-downloadurl' from closest element with the same 'href'
+    var href = $(this).attr('href');
+    var filename = $("div.attachments.dmsf_parent_container > p > a[href='" + href + "'].dmsf-icon-file").first().attr('data-downloadurl').split(':')[1];
+    // create 3 fancybox 'rel' groups to avoid image duplicates in slideshow
+    var relgroup = '';
+    if($(this).closest('div.thumbnails').length) {
+      relgroup = 'thumbnails';
+    } else if($(this).hasClass('icon-only')) {
+      relgroup = 'icon';
+    } else if($(this).hasClass('thumbnail')) {
+      relgroup = 'imagelink';
+    }
+    // Also support PDF preview in lightbox
+    var isPdf = filename.match(/\.pdf$/i);
+    // only apply thumbnail class to image and pdf links
+    if(filename.match(extensionRegexAll)) {
+      $(this)
+        .addClass('lightbox')
+        .attr('data-fancybox-type', isPdf ? 'iframe' : 'image')
+        .attr('title', filename)
+        .attr('rel', 'dmsf-' + relgroup);
+        // do not remove 'data-downloadurl' here otherwise the filename extraction crashes for following dmsf thumbnails
+    }
   });
 
-    // Add Fancybox to PDF links
-  $("div.attachments a.pdf," +
-    "div.journal ul.details a[href$='.pdf']:not(.icon-download)," +
-    "div.journal ul.details a[href$='.PDF']:not(.icon-download)," +
-    "div.journal div.thumbnails a[href$='.pdf']," +
-    "div.journal div.thumbnails a[href$='.PDF']").fancybox({
-      type          : 'iframe',
-      prevEffect    : 'none',
-      nextEffect    : 'none',
-      openSpeed     : 300,
-      closeSpeed    : 150,
-      width         : '90%',
-      height        : '90%',
-      autoSize      : true,
-      iframe : {
-        preload: false
-      }
+  // Add Fancybox to image links
+  $("div.attachments a.lightbox")
+  .add("div.attachments a.lightbox_preview")
+  .add("div.journal ul.details a:not(.icon-download)").filter((index,elem) => $(elem).attr('href').match(extensionRegexImage))
+  .add("div.journal div.thumbnails a")
+  .add("div.wiki a.thumbnail")
+  .add(".avatar a")
+  .fancybox({
+    prevEffect    : 'none',
+    nextEffect    : 'none',
+    openSpeed     : 300,
+    closeSpeed    : 150
+  });
+
+  // Add Fancybox to PDF links
+  $("div.attachments a.pdf")
+  .add("div.journal ul.details a:not(.icon-download)").filter((index,elem) => $(elem).attr('href').match(/\.pdf$/i))
+  .add("div.journal div.thumbnails a").filter((index,elem) => $(elem).attr('href').match(/\.pdf$/i))
+  .fancybox({
+    type          : 'iframe',
+    prevEffect    : 'none',
+    nextEffect    : 'none',
+    openSpeed     : 300,
+    closeSpeed    : 150,
+    width         : '90%',
+    height        : '90%',
+    autoSize      : true,
+    iframe : {
+      preload: false
+    }
   });
 });